Medical service creation updates.

+++ src/Pusula.Training.HealthCare.Blazor/Components/Pages/Patients.razor.cs
@@ -62,6 +62,7 @@ public partial class Patients
     protected override async Task OnInitializedAsync()
     {
         await SetPermissionsAsync();
+
     }
 
     protected override async Task OnAfterRenderAsync(bool firstRender)
@@ -176,8 +177,7 @@ public partial class Patients
     private async Task OpenEditPatientModalAsync(PatientDto input)
     {
         SelectedEditTab = "patient-edit-tab";
-
-
+        
         var patient = await PatientsAppService.GetAsync(input.Id);
 
         EditingPatientId = patient.Id;
@@ -189,10 +189,28 @@ public partial class Patients
 
     private async Task DeletePatientAsync(PatientDto input)
     {
+        
+        var confirmed = await UiMessageService.Confirm($"Are you sure you want to delete {input.FirstName} {input.LastName}?");
+        if(!confirmed) return;
+        
         await PatientsAppService.DeleteAsync(input.Id);
         await GetPatientsAsync();
     }
 
+    // revert delete function
+    private async Task RevertPatientAsync(PatientDto input)
+    {
+        
+        var confirmed = await UiMessageService.Confirm($"Are you sure you want to undelete {input.FirstName} {input.LastName}?");
+        if (!confirmed) return;
+        
+        var revertPatient = ObjectMapper.Map<PatientDto, PatientUpdateDto>(input);
+        revertPatient.IsDeleted = !input.IsDeleted;
+        await PatientsAppService.UpdateAsync(input.Id, revertPatient);
+        await GetPatientsAsync();
+    }
+
+    
     private async Task CreatePatientAsync()
     {
         try
@@ -235,7 +253,13 @@ public partial class Patients
             await HandleErrorAsync(ex);
         }
     }
-
+    
+    protected virtual async Task OnDeletedData(bool? isDeleted)
+    {
+        Filter.IsDeleted = isDeleted;
+        await SearchAsync();
+    }
+    
     protected virtual async Task OnFirstNameChangedAsync(string? firstName)
     {
         Filter.FirstName = firstName;
@@ -281,26 +305,9 @@ public partial class Patients
         Filter.MobilePhoneNumber = mobilePhoneNumber;
         await SearchAsync();
     }
-    
-    protected virtual async Task OnPatientTypeChangedAsync(EnumPatientTypes? selectedType)
+    protected virtual async Task OnGenderChangedAsync(int? gender)
     {
-        Filter.PatientType = selectedType;
-        await SearchAsync();
-    }
-    
-    protected virtual async Task OnInsuranceTypeChangedAsync(EnumInsuranceType? insuranceType)
-    {
-        Filter.InsuranceType = insuranceType;
-        await SearchAsync();
-    }
-    protected virtual async Task OnDiscountGroupChangedAsync(EnumDiscountGroup? discountGroup)
-    {
-        Filter.DiscountGroup = discountGroup;
-        await SearchAsync();
-    }
-    protected virtual async Task OnGenderChangedAsync(EnumGender? gender)
-    {
-        Filter.Gender = gender;
+        //Filter.Gender = gender;
         await SearchAsync();
     }
     private Task SelectAllItems()
